split integration tests and fix lint
