+++ .github/workflows/integration_test.yaml
@@ -24,7 +24,9 @@ jobs:
                 "test_http_interface.py",
                 "test_ingress.py",
                 "test_ingress_per_unit.py",
-                "test_haproxy_route_tcp.py"
+                "test_haproxy_route_tcp.py",
+                "test_haproxy_route_grpc.py",
+                "test_haproxy_route_https_backend.py"
               ]
           - name: haproxy-spoe-auth-operator
             working-directory: ./haproxy-spoe-auth-operator
